feat: Only Moderators AND the author can edit/delete

+++ src/Controller/ModeratorController.php
@@ -9,11 +9,11 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-// TODO: Make users also be able to edit and delete their own messages
-#[Route('/moderator')]
-#[IsGranted("ROLE_MODERATOR")]
+
+#[IsGranted("ROLE_USER")]
 class ModeratorController extends AbstractController
 {
     private $em;
@@ -24,6 +24,8 @@ class ModeratorController extends AbstractController
     #[Route('/message/edit/{id}', name: 'mod_edit_message')]
     public function edit_message(Message $message, Request $request): Response
     {
+        if (!$this->canEdit($message)) { return $this->redirectMessage($message); }
+
         // We fetch the form that we're either using, or need to use
         $form = $this->createForm(MessageType::class, $message);
         $form->handleRequest($request);
@@ -47,6 +49,8 @@ class ModeratorController extends AbstractController
     #[Route('/message/delete/{id}', name: 'mod_delete_message')]
     public function deleteMessage(Message $message) : Response
     {
+        if (!$this->canEdit($message)) { return $this->redirectMessage($message); }
+        
         $this->em->remove($message);
         $this->em->flush();
 
@@ -58,7 +62,7 @@ class ModeratorController extends AbstractController
 
     // Utility Functions
     // We need to check where to redirect the moderator
-    public function redirectMessage(Message $message) : Response
+    private function redirectMessage(Message $message) : Response
     {
         // If it has a subject, redirect to the subject in question
         $subject = $message->getSubject();
@@ -78,5 +82,13 @@ class ModeratorController extends AbstractController
 
         // If it has none, redirect to the homepage, still, something went wrong
         return $this->redirectToRoute("app_home");
+    }
+
+    // We check if the user is the author of the message
+    private function canEdit(Message $message): bool  {
+        $isAuthor = $message->getAuthor()->getId() == $this->getUser()->getId();
+        $isModerator = $this->isGranted('ROLE_MODERATOR');
+
+        return $isAuthor || $isModerator;
     }
 }
